@@ -13,54 +13,55 @@
   "private": true,
   "dependencies": {
     "@agm/core": "latest",
-    "@angular/animations": "^5.1.0",
-    "@angular/cdk": "^5.1.0",
-    "@angular/common": "^5.1.0",
-    "@angular/compiler": "^5.1.0",
-    "@angular/core": "^5.1.0",
-    "@angular/flex-layout": "2.0.0-beta.10",
-    "@angular/forms": "^5.1.0",
-    "@angular/http": "^5.1.0",
-    "@angular/material": "^5.1.0",
-    "@angular/material-moment-adapter": "^5.1.0",
-    "@angular/platform-browser": "^5.1.0",
-    "@angular/platform-browser-dynamic": "^5.1.0",
-    "@angular/router": "^5.1.0",
+    "@angular/animations": "^6.0.0",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/material": "^6.0.0",
+    "@angular/material-moment-adapter": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/router": "^6.0.0",
     "@google-cloud/functions-emulator": "^1.0.0-alpha.29",
-    "angularfire2": "latest",
+    "angularfire2": "^5.0.0-rc.5-next",
     "bootstrap": "3.3.7",
     "core-js": "^2.4.1",
-    "firebase": "^4.3.1",
+    "firebase": "^4.13.1",
     "firebaseui": "2.5.1",
-    "jspdf": "1.3.5",
     "hammerjs": "2.0.8",
+    "jspdf": "1.3.5",
     "lodash": "latest",
     "moment": "latest",
     "ngx-bootstrap": "1.7.1",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/core": "latest",
-    "@angular/cli": "^1.6.3",
-    "@angular/compiler-cli": "^5.1.0",
-    "@angular/language-service": "^5.1.0",
-    "@types/jasmine": "~2.5.53",
+    "@angular/cli": "^6.0.7",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/language-service": "^6.0.0",
+    "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/lodash": "latest",
     "@types/node": "~6.0.60",
     "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "ts-node": "~4.1.0",
+    "tslint": "~5.9.1",
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.0"
   }
 }
