@@ -1,6 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Observable} from "rxjs/Observable";
-import {Subject} from "rxjs/Subject";
+import {Observable, Subject} from "rxjs";
 
 @Injectable()
 export class ImageService {
