@@ -1,8 +1,10 @@
+
+import {refCount, publishReplay, map} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {AngularFirestore, DocumentChangeAction} from 'angularfire2/firestore';
-import 'rxjs/add/operator/publishReplay';
-import 'rxjs/add/operator/map';
+
+
 import {FootballCamp} from '../../models/football-camp';
 
 @Injectable()
@@ -17,28 +19,28 @@ export class FootballCampService {
     if (this.footballCamps$ == null) {
       this.footballCamps$ = this.angularFirestore
         .collection<FootballCamp>('camps')
-        .snapshotChanges()
-        .map<DocumentChangeAction[], FootballCamp[]>(actions => {
+        .snapshotChanges().pipe(
+        map<DocumentChangeAction[], FootballCamp[]>(actions => {
           return actions.map(action => {
             const data = action.payload.doc.data() as FootballCamp;
             data.id = action.payload.doc.id;
             return data as FootballCamp;
           });
-        })
-        .publishReplay(1) // Latest event is buffered and will be emit to new subscriber
-        .refCount(); // Transform ConnectableObservable to Observable and handle multiple subscription / unsubscription
+        }),
+        publishReplay(1), // Latest event is buffered and will be emit to new subscriber
+        refCount(), ); // Transform ConnectableObservable to Observable and handle multiple subscription / unsubscription
     }
 
     return this.footballCamps$;
   }
 
   getFootballCamp(campId: string): Observable<FootballCamp> {
-    return this.getFootballCamps()
-      .map<FootballCamp[], FootballCamp>((footballCamps: FootballCamp[]) => {
+    return this.getFootballCamps().pipe(
+      map<FootballCamp[], FootballCamp>((footballCamps: FootballCamp[]) => {
           return footballCamps.find(footballCamp => {
             return campId === footballCamp.id;
           });
         }
-      );
+      ));
   }
 }
