@@ -1,6 +1,7 @@
+import {publishReplay, map, refCount} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
 import {AngularFirestore, DocumentChangeAction} from 'angularfire2/firestore';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {Registration} from '../../models/registration';
 
 @Injectable()
@@ -28,15 +29,16 @@ export class RegistrationService {
       .collection<Registration>('registrations', ref => {
         return ref.where('sessionId', '==', sessionId)
       })
-      .snapshotChanges()
-      .map<DocumentChangeAction[], Registration[]>(actions => {
-        return actions.map(action => {
-          const data = action.payload.doc.data() as Registration;
-          data.id = action.payload.doc.id;
-          return data as Registration;
-        });
-      })
-      .publishReplay(1) // Latest event is buffered and will be emit to new subscriber
-      .refCount(); // Transform ConnectableObservable to Observable and handle multiple subscription / unsubscription
+      .snapshotChanges().pipe(
+        map<DocumentChangeAction[], Registration[]>(actions => {
+          return actions.map(action => {
+            const data = action.payload.doc.data() as Registration;
+            data.id = action.payload.doc.id;
+            return data as Registration;
+          });
+        }),
+        publishReplay(1), // Latest event is buffered and will be emit to new subscriber
+        refCount() // Transform ConnectableObservable to Observable and handle multiple subscription / unsubscription
+      )
   }
 }
